Move fetching photos to viewModel
