+++ CodingSession/GalleryViewModel.swift
@@ -1,0 +1,28 @@
+//
+//  GalleryViewModel.swift
+//  CodingSession
+//
+//  Created by Andrii Kvashuk on 13/05/2024.
+//
+
+import Foundation
+import RxSwift
+import Photos
+
+class GalleryViewModel {
+    
+    func fetchPhotos() -> [PHAsset] {
+        let fetchOptions = PHFetchOptions()
+        fetchOptions.predicate = NSPredicate(format: "mediaType == %d", PHAssetMediaType.video.rawValue)
+        
+        let fetchResult = PHAsset.fetchAssets(with: fetchOptions)
+        var videoAssets: [PHAsset] = []
+           
+        fetchResult.enumerateObjects { (asset, _, _) in
+            videoAssets.append(asset)
+        }
+        
+        return videoAssets
+    }
+    
+}
